% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heightmap.R
\name{create_heightmap}
\alias{create_heightmap}
\title{Heightmap creation}
\usage{
create_heightmap(rad, min = 1, max = 255)
}
\arguments{
\item{rad}{An integer value, will be an exponent of 2. It is not recommended to set this higher than 10.}

\item{min}{A numeric of minimum value for a cell/pixel (default 1)}

\item{min}{A numeric of maximum value for a cell/pixel (default 255)}
}
\value{
A matrix with random integers between min and max in corners and value -1 in every other field.
}
\description{
Create an empty square matrix with size 2^x + 1
}
