% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_step.R
\name{diamond_step}
\alias{diamond_step}
\title{Diamond Step}
\usage{
diamond_step(HM, steps)
}
\arguments{
\item{HM}{A matrix containing the heightmap that is being generated}

\item{steps}{A vector of string containing planned steps for the diamond function in the format "row,column,radius"}
}
\value{
Returns an altered Heightmap entered as parameter
}
\description{
Execute diamond step of the diamond square algorithm
}
