% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pop}
\alias{pop}
\title{Stack functionality}
\usage{
pop(x)
}
\arguments{
\item{x}{A vector to pop an item out of}
}
\value{
The last item of vector x
}
\description{
Pop operation on a vector. Edits input vector and returns the item that was pushed in most recently
}
