% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{push}
\alias{push}
\title{Stack functionality}
\usage{
push(x, value)
}
\arguments{
\item{x}{A vector to push an item into}

\item{value}{A value to push into the stack (at the end of the vector)}
}
\description{
Push operation on a vector.
}
