% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_step.R
\name{square_step}
\alias{square_step}
\title{Square Step}
\usage{
square_step(HM, steps)
}
\arguments{
\item{HM}{A matrix containing the heightmap that is being generated}

\item{steps}{A vector containing planned steps for the square function in the format "row,column,radius"}
}
\value{
Returns an altered Heightmap entered as parameter
}
\description{
Execute square step of the diamond square algorithm
}
